Move type casters into files that define their types
Since we're using pybind11 everywhere, it should be fine now to access
it in any header, and putting the type caster there is clearer. We don't
need the weird macro checks to conditionally define them either.
